Renamed the LogManager namespace to the LogCategories.
