+++ CMakeLists.txt
@@ -123,7 +123,7 @@ endif()
 # ======================================================
 
 set(LIBRARY_HEADERS
-    src/lib/logmanager/logmanager.h
+    src/lib/logcategories/logcategories.h
     src/lib/tunedtypes/tunedtypes.h
     src/lib/translationmanager/translationmanager.h
     src/lib/notificationsmanager/dbusconverters.h
@@ -132,7 +132,7 @@ set(LIBRARY_HEADERS
 )
 
 set(LIBRARY_SOURCES
-    src/lib/logmanager/logmanager.cpp
+    src/lib/logcategories/logcategories.cpp
     src/lib/tunedtypes/tunedtypes.cpp
     src/lib/translationmanager/translationmanager.cpp
     src/lib/notificationsmanager/notificationsmanager.cpp
