+++ src/lib/logcategories/logcategories.cpp
@@ -5,15 +5,15 @@
 */
 
 /**
- * @file logmanager/logmanager.cpp
- * Contains the LogManager namespace implementation.
+ * @file logcategories/logcategories.cpp
+ * Contains the LogCategories namespace implementation.
 */
 
 #include <QLoggingCategory>
 
-#include "logmanager/logmanager.h"
+#include "logcategories/logcategories.h"
 
-namespace LogManager
+namespace LogCategories
 {
     Q_LOGGING_CATEGORY(DBus, "tuned-switcher.dbus")
 }
